requirements.txt
python-docx
PyPDF2
pdfplumber
spacy
nltk
sentence-transformers
google-generativeai

